lazy-load admin routes; full-width + sticky-thead /admin/routes
Every admin route now loads on demand:
- Pure-component routes were renamed to *.lazy.tsx and switched to
  createLazyFileRoute: admin, $lang.phrases.index, $lang.phrases.$id,
  $lang.requests.index, $lang.requests.$id.
- Routes that need staticData / loader (admin/index, admin/routes,
  $lang, $lang.phrases, $lang.requests) are now split into an eager
  *.tsx file (route metadata only) plus a *.lazy.tsx sibling with the
  component.

New staticData flag: fullWidth.
- _user.tsx layout: removes the max-w cap (was 4xl, 6xl with
  wideContent) when fullWidth is set.
- /admin/routes uses fullWidth so the introspection table can spread.

/admin/routes table polish:
- thead is sticky to the top of the scroll container, with a
  background and shadow so rows don't show through.
- The lazy column is now a clearer "lazy" / "eager" pill instead of a
  ✓ / · glyph.
- Added a "full" column for fullWidth.
