+++ src/routes/_user/admin.lazy.tsx
@@ -1,0 +1,5 @@
+import { createLazyFileRoute, Outlet } from '@tanstack/react-router'
+
+export const Route = createLazyFileRoute('/_user/admin')({
+	component: () => <Outlet />,
+})
