--- src/routes/_user/admin.tsx
@@ -1,5 +1,0 @@
-import { createFileRoute, Outlet } from '@tanstack/react-router'
-
-export const Route = createFileRoute('/_user/admin')({
-	component: () => <Outlet />,
-})
