+++ src/routes/_user/admin/$lang.lazy.tsx
@@ -1,0 +1,5 @@
+import { createLazyFileRoute, Outlet } from '@tanstack/react-router'
+
+export const Route = createLazyFileRoute('/_user/admin/$lang')({
+	component: () => <Outlet />,
+})
