+++ src/routes/_user/admin/$lang.requests.lazy.tsx
@@ -1,0 +1,31 @@
+import { CSSProperties } from 'react'
+import { createLazyFileRoute, Outlet } from '@tanstack/react-router'
+import { useAuth } from '@/lib/use-auth'
+import Callout from '@/components/ui/callout'
+import { ShieldAlert } from 'lucide-react'
+
+export const Route = createLazyFileRoute('/_user/admin/$lang/requests')({
+	component: AdminRequestsLayout,
+})
+
+const style = { viewTransitionName: 'main-area' } as CSSProperties
+
+function AdminRequestsLayout() {
+	const { isAdmin, userEmail } = useAuth()
+
+	return (
+		<main style={style} data-testid="admin-requests-page">
+			{!isAdmin && (
+				<div data-testid="admin-not-authorized-warning">
+					<Callout variant="problem" Icon={ShieldAlert}>
+						<p>
+							You are logged in as <strong>{userEmail}</strong> who is not an
+							admin user; the forms on this page will not work.
+						</p>
+					</Callout>
+				</div>
+			)}
+			<Outlet />
+		</main>
+	)
+}
