+++ src/routes/_user/admin/$lang.tsx
@@ -1,11 +1,10 @@
-import { createFileRoute, Outlet } from '@tanstack/react-router'
+import { createFileRoute } from '@tanstack/react-router'
 import { phrasesCollection } from '@/features/phrases/collections'
 import { phraseRequestsCollection } from '@/features/requests/collections'
 import { langTagsCollection } from '@/features/languages/collections'
 import { publicProfilesCollection } from '@/features/profile/collections'
 
 export const Route = createFileRoute('/_user/admin/$lang')({
-	component: () => <Outlet />,
 	loader: async () => {
 		await Promise.all([
 			phrasesCollection.preload(),
